Refactor and various fixes:
- Enable the export of banks with 1bpp for the binary version.
- Fixed bank exporting when it contains sprites with custom sizes.
- Fixed get color from the bank when the sprites were custom size.
- Now the palette generation logic is in a common place to be called by
anyone.
- Now the bank generation wont be cancel if the palette is not found in
a character. It will just generate a temporal palette based on what the
bank has.
- Fixed a crash when generating a palette for an 8bpp that required a
folder creation.

+++ GBATool/Utils/Image/ImageProcessing.cs
@@ -112,7 +112,7 @@ public static class ImageProcessing
     {
         int outputIndex = pixelIndex / 8;
 
-        byte store = (byte)((byte)colorIndex << (pixelIndex % 8));
+        byte store = (byte)(colorIndex << (pixelIndex % 8));
 
         output[outputIndex] |= store;
     }
